[Window/Core][Window Wrapper] Moved window from core to it's separate module
