+++ Engine/Modules/CMakeLists.txt
@@ -1,5 +1,6 @@
 cmake_minimum_required(VERSION 3.20)
 
 add_subdirectory(Core)
+add_subdirectory(Window)
 add_subdirectory(Rhi)
 add_subdirectory(Rpi)
