+++ Engine/Modules/Core/CMakeLists.txt
@@ -31,8 +31,6 @@ target_include_directories(SynergonCore PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}/Include
 )
 
-target_link_libraries(SynergonCore PUBLIC glfw)
-
 set_property(TARGET SynergonCore PROPERTY FOLDER "Engine/Modules/Core")
 
 add_subdirectory(Tests)
